Refactor property deduplication and enhance image caching
- Updated `get_recent_properties_for_dedup` to accept `None` for loading all properties.
- Added `run_dedup_existing` function to retroactively merge duplicate properties in the database.
- Implemented `copy_cached_images` and `find_cached_file` functions for improved image handling.
- Modified image URL resolution in routes to support cached images regardless of index.
- Enhanced CSS styles for better UI presentation.
- Added integration tests for the new deduplication functionality and image caching.
- Updated various filters and quality checks to accommodate new data types.

+++ src/home_finder/filters/quality.py
@@ -1093,10 +1093,10 @@ class PropertyQualityFilter:
         effective_average: int,
         area_context: str | None,
         outcode: str | None,
-        council_tax_band_c: int | None,
+        council_tax_band_c: float | None,
         crime_summary: str | None,
         rent_trend: str | None,
-        energy_estimate: int | None,
+        energy_estimate: float | None,
         hosting_tolerance: str | None,
     ) -> dict[str, Any]:
         """Run Phase 2 evaluation API call.
@@ -1284,8 +1284,8 @@ class PropertyQualityFilter:
         features: list[str] | None = None,
         area_context: str | None = None,
         outcode: str | None = None,
-        council_tax_band_c: int | None = None,
-        energy_estimate: int | None = None,
+        council_tax_band_c: float | None = None,
+        energy_estimate: float | None = None,
         crime_summary: str | None = None,
         rent_trend: str | None = None,
         hosting_tolerance: str | None = None,
